+++ pie-custom-functions-mu.php
@@ -42,6 +42,7 @@ if ( version_compare( PHP_VERSION, '8.0', '<' ) ) {
 include_once plugin_dir_path( __FILE__ ) . 'pie/redirections.php';
 include_once plugin_dir_path( __FILE__ ) . 'pie/utilities.php';
 include_once plugin_dir_path( __FILE__ ) . 'pie/branda-config.php';
+include_once plugin_dir_path( __FILE__ ) . 'pie/security-headers.php';
 
 pie_hide_plugin( 'pie-custom-functions/pie-custom-functions.php' );