+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] 2026-06-15 
 - Add shared functions file 
 - Update hide plugins and include hummingbird plugin for non-Pie users
 - Add Security Headers functionality